Tests/LibWeb/Text/input/css/small-aspect-ratio.html

<!DOCTYPE html>
<script src="../include.js"></script>
<script>
    test(() => {
        const smallAspectRatios = [
            ".0000000000001 / .00000000000001",
            "1/0.00000000000001",
            "0.00000000000001/1",
        ];
        for (const ratio of smallAspectRatios) {
            const element = document.createElement("div");
            element.style.width = "100px";
            element.style.aspectRatio = ratio;
            document.body.appendChild(element);
            println(`element height: ${element.clientHeight}px`);
            element.remove();
        }
    });
</script>
