Tests/LibWeb/Text/expected/css/small-aspect-ratio.txt

element height: 10px
element height: 0px
element height: 0px
